获取链接错误码接口&更新RESET_STREAM,CONNECTION_CLOSE,GO_AWAY 3种帧格式

+++ transport/xqc_frame_parser.h
@@ -38,21 +38,21 @@ int xqc_gen_ack_frame(xqc_connection_t *conn, xqc_packet_out_t *packet_out, xqc_
 
 int xqc_parse_ack_frame(xqc_packet_in_t *packet_in, xqc_connection_t *conn, xqc_ack_info_t *ack_info);
 
-int xqc_gen_conn_close_frame(xqc_packet_out_t *packet_out, unsigned short err_code, int is_app, int frame_type);
+int xqc_gen_conn_close_frame(xqc_packet_out_t *packet_out, uint64_t err_code, int is_app, int frame_type);
 
-int xqc_parse_conn_close_frame(xqc_packet_in_t *packet_in, unsigned short *err_code);
+int xqc_parse_conn_close_frame(xqc_packet_in_t *packet_in, uint64_t *err_code);
 
 int xqc_gen_reset_stream_frame(xqc_packet_out_t *packet_out, xqc_stream_id_t stream_id,
-                           unsigned short err_code, uint64_t final_size);
+                           uint64_t err_code, uint64_t final_size);
 
 int xqc_parse_reset_stream_frame(xqc_packet_in_t *packet_in, xqc_stream_id_t *stream_id,
-                             unsigned short *err_code, uint64_t *final_size);
+                             uint64_t *err_code, uint64_t *final_size);
 
 int xqc_gen_stop_sending_frame(xqc_packet_out_t *packet_out, xqc_stream_id_t stream_id,
-                               unsigned short err_code);
+                               uint64_t err_code);
 
 int xqc_parse_stop_sending_frame(xqc_packet_in_t *packet_in, xqc_stream_id_t *stream_id,
-                                 unsigned short *err_code);
+                                 uint64_t *err_code);
 
 int xqc_gen_data_blocked_frame(xqc_packet_out_t *packet_out, uint64_t data_limit);
